split the huge the matherinfo.go file into smaller files
